+++ src/tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod compiler;
+pub mod text_process;
